Clean up; Remove DOM Editor tab; Add DOM form to Changes tab
