+++ src/visual_editor/VisualEditorSection.tsx
@@ -19,7 +19,7 @@ const VisualEditorSection: FC<{
   return (
     <>
       <div
-        className={clsx("text-slate-300 text-xs font-semibold uppercase p-4", {
+        className={clsx("text-light text-xs font-semibold uppercase p-4", {
           "flex justify-between": onClose,
           "shadow-xl": isExpanded,
           "mb-2": isExpanded,
@@ -29,22 +29,16 @@ const VisualEditorSection: FC<{
           {title}
           {isCollapsible ? (
             <RxCaretDown
-              className={clsx(
-                "w-4 h-4 cursor-pointer hover:text-slate-100 mx-2",
-                {
-                  "rotate-180": !isExpanded,
-                }
-              )}
+              className={clsx("w-4 h-4 cursor-pointer text-link mx-2", {
+                "rotate-180": !isExpanded,
+              })}
               onClick={() => setIsExpanded(!isExpanded)}
             />
           ) : null}
         </div>
 
         {onClose && (
-          <button
-            className="text-slate-300 hover:text-slate-100"
-            onClick={onClose}
-          >
+          <button className="text-link" onClick={onClose}>
             <RxCross1 className="w-3 h-3" />
           </button>
         )}
